backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

faster-whisper==1.0.3
whisperx==3.1.1
pyannote.audio==3.1.1
onnxruntime-gpu==1.19.0

# Translation stack (supports tc-big ko->en and alt models via env)
transformers==4.44.2
sentencepiece==0.2.0
huggingface_hub==0.24.6

TTS==0.22.0
soundfile==0.12.1

sacremoses==0.0.53
cutlet==0.4.0
fugashi==1.3.2
unidic-lite==1.0.8
torchcodec

# Worker dependencies
boto3==1.34.160
requests==2.32.3
